feat: auth error provider
