+++ src/components/auth/social-sign-in.tsx
@@ -1,21 +1,13 @@
 "use client";
 
-import { useRouter, useSearchParams } from "next/navigation";
 import { GoogleIcon } from "../icons/google-icon";
 import { Button } from "../ui/button";
 import { signIn } from "next-auth/react";
 import { toast } from "../ui/use-toast";
-import { useEffect } from "react";
 import { useMutation } from "@tanstack/react-query";
 import { RotateCwIcon } from "lucide-react";
-import { paths } from "@/routes/paths";
 
 export const SocialSignInForm = () => {
-  const searchParams = useSearchParams();
-  const error = searchParams.get("error");
-
-  const router = useRouter();
-
   const { mutate, isLoading } = useMutation({
     mutationFn: () => signIn("google"),
     onError: () =>
@@ -24,19 +16,6 @@ export const SocialSignInForm = () => {
         description: "Failed to sign out. Please try again.",
       }),
   });
-
-  useEffect(() => {
-    if (error && error === "OAuthAccountNotLinked") {
-      toast({
-        variant: "destructive",
-        title: "Email Already in Use",
-        description:
-          "This email is already associated with an account. Please try logging in using this email or reach out to our support team for assistance. Thank you!",
-      });
-
-      router.replace(paths.auth.signIn);
-    }
-  }, [error, router]);
 
   return (
     <Button
